IndicoIo-Python/requirements.txt
six>=1.3.0
requests>=2.2.1
Pillow>=2.8.2
mock>=1.3.0
futures>=3.0.0
